+++ changelog.md
@@ -7,6 +7,7 @@
 
 - Add a new `--no-config` option to disable loading of configuration files. Closes {issue}`750`.
 - Remove the `-C` short option associated with `--config`.
+- Remove the `-t` short option associated with `--table-format`.
 - Fix display in `--show-params` of parameters sharing the same name but in different commands.
 - Replace unmaintained `mergedeep` dependency by `deepmerge`.
 - Expose `click._utils.UNSET` sentinel as `click_extra.UNSET`.